// src/app/models/customer.model.ts
export interface Customer {
  customerId?: number;
  customerName: string;
  residence?: string;
  description?: string;
  job?: string;
  enteredBy?: string;
  enteredDate?: Date;
  lastModifiedBy?: string;
  lastModifiedDate?: Date;
  customerSource?: string;
  customerClassification?: string;
  customerAddress?: string;
  firstPhone?: string;
  secondaryPhone?: string;
  whatsApp?: string;
  email?: string;
  customerCode?: string;
  nationality?: string;
  salesManName?: string;
  salesmanId?: number;  
  }